move XsecBiasingOperator base class into BiasOperators library
