+++ SimCore/include/SimCore/BiasOperators/GammaToMuPair.h
@@ -1,7 +1,7 @@
 #ifndef SIMCORE_BIASOPERATORS_GAMMATOMUPAIR_H_
 #define SIMCORE_BIASOPERATORS_GAMMATOMUPAIR_H_
 
-#include "SimCore/XsecBiasingOperator.h"
+#include "SimCore/BiasOperators/XsecBiasingOperator.h"
 
 namespace simcore {
 namespace biasoperators {
